Scope CF.INSERT filter-full (-1) assertion to RESP2
The RedisBloom server returns the per-item -1 (filter full) as an integer
only over RESP2; over RESP3 it returns boolean false instead, so the -1 vs 0
distinction is not observable on the client. Move the -1 -> null integration
assertion into RedisCuckooFilterResp2IntegrationTests and drop the filter-full
case from the RESP3 sync/reactive tests. The output mapping is independently
covered by CuckooInsertBooleanListOutput/ValueListOutput unit tests.
